prompt-sandbox/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
ariadne==0.17.1
asn1crypto==1.5.1
astroid==2.14.2
async-timeout==4.0.2
attrs==22.1.0
black==23.1.0
blinker==1.5
boto3==1.26.72
botocore==1.29.72
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coverage==7.1.0
cryptography==38.0.4
debugpy==1.6.6
dill==0.3.6
docutils==0.19
filelock==3.8.2
Flask==2.2.3
frozenlist==1.3.3
graphql-core==3.2.3
idna==3.4
iniconfig==1.1.1
isort==5.11.2
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
lazy-object-proxy==1.8.0
localstack-client==1.39
MarkupSafe==2.1.1
mccabe==0.7.0
moto==4.1.3
multidict==6.0.4
mypy==1.0.0
mypy-extensions==0.4.3
numpy==1.24.2
openai==0.27.7
oscrypto==1.3.0
packaging==22.0
pandas==1.5.3
pathspec==0.10.3
pip-tools==6.12.2
platformdirs==2.6.0
pluggy==1.0.0
pycparser==2.21
pycryptodomex==3.16.0
PyJWT==2.6.0
pylint==2.16.2
pyOpenSSL==22.1.0
pyproject_hooks==1.0.0
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.6
requests==2.28.1
responses==0.22.0
s3transfer==0.6.0
scikit-learn==1.2.0
scipy==1.9.3
sentry-sdk==1.15.0
six==1.16.0
sniffio==1.3.0
snowflake-connector-python==3.0.0
starlette==0.20.4
terraform-local==0.6
threadpoolctl==3.1.0
toml==0.10.2
tomlkit==0.11.6
tqdm==4.65.0
types-toml==0.10.8.5
typing_extensions==4.4.0
urllib3==1.26.13
waitress==2.1.2
Werkzeug==2.2.2
wrapt==1.14.1
xmltodict==0.13.0
yarl==1.9.2
